Add CMakeLists files for accumulate_cut_c
