--- _LowLevelCode/cpp/accumulate_cut_c/accumulate_cut_c.sln
@@ -1,26 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 10.00
-# Visual Studio 2008
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "accumulate_cut_c", "accumulate_cut_c\accumulate_cut_c.vcproj", "{FFD6E603-ED85-4A3A-A35E-6D847CE2AB19}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Debug|x64 = Debug|x64
-		Release|Win32 = Release|Win32
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FFD6E603-ED85-4A3A-A35E-6D847CE2AB19}.Debug|Win32.ActiveCfg = Debug|Win32
-		{FFD6E603-ED85-4A3A-A35E-6D847CE2AB19}.Debug|Win32.Build.0 = Debug|Win32
-		{FFD6E603-ED85-4A3A-A35E-6D847CE2AB19}.Debug|x64.ActiveCfg = Debug|x64
-		{FFD6E603-ED85-4A3A-A35E-6D847CE2AB19}.Debug|x64.Build.0 = Debug|x64
-		{FFD6E603-ED85-4A3A-A35E-6D847CE2AB19}.Release|Win32.ActiveCfg = Release|Win32
-		{FFD6E603-ED85-4A3A-A35E-6D847CE2AB19}.Release|Win32.Build.0 = Release|Win32
-		{FFD6E603-ED85-4A3A-A35E-6D847CE2AB19}.Release|x64.ActiveCfg = Release|x64
-		{FFD6E603-ED85-4A3A-A35E-6D847CE2AB19}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
