--- _LowLevelCode/cpp/accumulate_cut_c/accumulate_cut_c/linkoptions.def
@@ -1,2 +1,0 @@
-EXPORTS
-  mexFunction
